Future-proof external crate compatibility
