+++ rkyv/Cargo.toml
@@ -31,17 +31,18 @@ rkyv_derive.workspace = true
 # Before adding support for another crate, please consider getting rkyv support
 # in the crate instead.
 
-arrayvec = { version = "0.7", optional = true, default-features = false }
-bytes = { version = "1.4.0", optional = true, default-features = false }
-# hashbrown
-indexmap = { version = "2.2", optional = true, default-features = false }
-smallvec = { version = "1.7", optional = true, default-features = false }
-smol_str_02 = { version = "0.2", package = "smol_str", optional = true, default-features = false }
-smol_str_03 = { version = "0.3", package = "smol_str", optional = true, default-features = false }
-thin-vec = { version = "0.2.12", optional = true, default-features = false }
-tinyvec = { version = "1.5", optional = true, default-features = false }
-triomphe = { version = "0.1", optional = true, default-features = false }
-uuid = { version = "1.3", optional = true, default-features = false }
+arrayvec-0_7 = { package = "arrayvec", version = "0.7", optional = true, default-features = false }
+bytes-1 = { package = "bytes", version = "1", optional = true, default-features = false }
+# rkyv already depends on hashbrown 0.14, so we can't duplicate this, but we can expose it as a feature below
+# hashbrown-0_14 = { package = "hashbrown", version = "0.14", optional = true, default-features = false }
+indexmap-2 = { package = "indexmap", version = "2", optional = true, default-features = false }
+smallvec-1 = { package = "smallvec", version = "1", optional = true, default-features = false }
+smol_str-0_2 = { package = "smol_str", version = "0.2", optional = true, default-features = false }
+smol_str-0_3 = { package = "smol_str", version = "0.3", optional = true, default-features = false }
+thin-vec-0_2 = { package = "thin-vec", version = "0.2.12", optional = true, default-features = false }
+tinyvec-1 = { package = "tinyvec", version = "1", optional = true, default-features = false }
+triomphe-0_1 = { package = "triomphe", version = "0.1", optional = true, default-features = false }
+uuid-1 = { package = "uuid", version = "1", optional = true, default-features = false }
 
 [features]
 default = ["std", "bytecheck"]
@@ -52,15 +53,15 @@ unaligned = []
 pointer_width_16 = []
 pointer_width_32 = []
 pointer_width_64 = []
-alloc = ["dep:hashbrown", "tinyvec?/alloc", "rancor/alloc"]
-std = ["alloc", "bytes?/std", "indexmap?/std", "ptr_meta/std", "uuid?/std"]
+alloc = ["dep:hashbrown", "tinyvec-1?/alloc", "rancor/alloc"]
+std = ["alloc", "bytes-1?/std", "indexmap-2?/std", "ptr_meta/std", "uuid-1?/std"]
 bytecheck = ["dep:bytecheck", "rend/bytecheck", "rkyv_derive/bytecheck"]
 
 # External crate support
-hashbrown = ["dep:hashbrown"]
-indexmap = ["dep:indexmap", "alloc"]
-triomphe = ["dep:triomphe", "alloc"]
-uuid = ["dep:uuid", "bytecheck?/uuid"]
+hashbrown-0_14 = ["dep:hashbrown"]
+indexmap-2 = ["dep:indexmap-2", "alloc"]
+triomphe-0_1 = ["dep:triomphe-0_1", "alloc"]
+uuid-1 = ["dep:uuid-1", "bytecheck?/uuid"]
 
 [package.metadata.docs.rs]
 features = ["bytecheck"]
