+++ rkyv/src/impls/ext/hashbrown_0_14/mod.rs
@@ -1,0 +1,6 @@
+mod hash_map;
+mod hash_set;
+mod with;
+
+// NOTE: When hashbrown updates to 0.15 or later,
+// this module will need to be updated to use hashbrown_0_14
