+++ rkyv/src/impls/ext/thin_vec_0_2.rs
@@ -1,5 +1,5 @@
-use arrayvec::ArrayVec;
 use rancor::Fallible;
+use thin_vec_0_2::ThinVec;
 
 use crate::{
     ser::{Allocator, Writer},
@@ -7,7 +7,7 @@ use crate::{
     Archive, Archived, Deserialize, Place, Serialize,
 };
 
-impl<T, const CAP: usize> Archive for ArrayVec<T, CAP>
+impl<T> Archive for ThinVec<T>
 where
     T: Archive,
 {
@@ -19,10 +19,10 @@ where
     }
 }
 
-impl<T, S, const CAP: usize> Serialize<S> for ArrayVec<T, CAP>
+impl<T, S> Serialize<S> for ThinVec<T>
 where
     T: Serialize<S>,
-    S: Fallible + Allocator + Writer + ?Sized,
+    S: Allocator + Writer + Fallible + ?Sized,
 {
     fn serialize(
         &self,
@@ -32,18 +32,17 @@ where
     }
 }
 
-impl<T, D, const CAP: usize> Deserialize<ArrayVec<T, CAP>, D>
+impl<T, D: Fallible + ?Sized> Deserialize<ThinVec<T>, D>
     for ArchivedVec<Archived<T>>
 where
     T: Archive,
     Archived<T>: Deserialize<T, D>,
-    D: Fallible + ?Sized,
 {
     fn deserialize(
         &self,
         deserializer: &mut D,
-    ) -> Result<ArrayVec<T, CAP>, D::Error> {
-        let mut result = ArrayVec::new();
+    ) -> Result<ThinVec<T>, D::Error> {
+        let mut result = ThinVec::with_capacity(self.len());
         for item in self.as_slice() {
             result.push(item.deserialize(deserializer)?);
         }
@@ -51,23 +50,20 @@ where
     }
 }
 
-impl<T, U, const CAP: usize> PartialEq<ArrayVec<U, CAP>> for ArchivedVec<T>
+impl<T, U> PartialEq<ThinVec<U>> for ArchivedVec<T>
 where
     T: PartialEq<U>,
 {
-    fn eq(&self, other: &ArrayVec<U, CAP>) -> bool {
+    fn eq(&self, other: &ThinVec<U>) -> bool {
         self.as_slice().eq(other.as_slice())
     }
 }
 
-impl<T, U, const CAP: usize> PartialOrd<ArrayVec<U, CAP>> for ArchivedVec<T>
+impl<T, U> PartialOrd<ThinVec<U>> for ArchivedVec<T>
 where
     T: PartialOrd<U>,
 {
-    fn partial_cmp(
-        &self,
-        other: &ArrayVec<U, CAP>,
-    ) -> Option<::core::cmp::Ordering> {
+    fn partial_cmp(&self, other: &ThinVec<U>) -> Option<::core::cmp::Ordering> {
         crate::impls::lexicographical_partial_ord(
             self.as_slice(),
             other.as_slice(),
@@ -77,14 +73,30 @@ where
 
 #[cfg(test)]
 mod tests {
-    use arrayvec::ArrayVec;
-
+    use super::ThinVec;
     use crate::api::test::roundtrip_with;
 
     #[test]
-    fn roundtrip_array_vec() {
-        roundtrip_with(&ArrayVec::<i32, 4>::from([10, 20, 40, 80]), |a, b| {
+    fn roundtrip_thin_vec() {
+        roundtrip_with(&ThinVec::<i32>::from_iter([10, 20, 40, 80]), |a, b| {
             assert_eq!(**a, **b)
         });
+    }
+
+    #[test]
+    fn test_partial_eq() {
+        use crate::Archive;
+
+        #[derive(Archive)]
+        #[rkyv(crate, compare(PartialEq, PartialOrd))]
+        struct Inner {
+            a: i32,
+        }
+
+        #[derive(Archive)]
+        #[rkyv(crate, compare(PartialEq, PartialOrd))]
+        struct Outer {
+            a: ThinVec<Inner>,
+        }
     }
 }
